FIX: Downloading via format chooser doesn't create $tmpfile in download.php. [t35012]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@23448 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/download.php
@@ -480,7 +480,7 @@ if(!hook('replacefileoutput'))
 
 // Deleting Exiftool temp File:
 // Note: Only for downloads (not previews)
-if(!$noattach && -1 == $alternative && $exiftool_write && file_exists($tmpfile))
+if(!$noattach && -1 == $alternative && $exiftool_write && isset($tmpfile) && file_exists($tmpfile))
     {
     delete_exif_tmpfile($tmpfile);
     }
